+++ src/app/home/home.module.ts
@@ -17,7 +17,7 @@ import { UpgradeComponent } from '../upgrade/upgrade.component';
 import { SharedComponentsModule } from '../shared/shared-components.module';
 import { UsersAchievementsModule } from '../users/users-achievements/users-achievements.module';
 import { NewsModule } from '../news/news.module';
-import { LogsMyPlanetComponent } from '../manager-dashboard/reports/myplanet/logs-myplanet.component';
+import { LogsMyPlanetComponent } from '../manager-dashboard/reports/myplanet/logs.component';
 import { TeamsModule } from '../teams/teams.module';
 import { CommunityComponent } from '../community/community.component';
 import { PlanetCalendarModule } from '../shared/calendar.module';
