+++ src/app/manager-dashboard/manager-dashboard-router.module.ts
@@ -9,8 +9,8 @@ import { ManagerDashboardConfigurationComponent } from './manager-dashboard-conf
 import { ReportsComponent } from './reports/reports.component';
 import { ReportsDetailComponent } from './reports/reports-detail.component';
 import { ReportsPendingComponent } from './reports/reports-pending.component';
-import { ReportsMyPlanetComponent } from './reports/myplanet/reports-myplanet.component';
-import { LogsMyPlanetComponent } from './reports/myplanet/logs-myplanet.component';
+import { ReportsMyPlanetComponent } from './reports/myplanet/reports.component';
+import { LogsMyPlanetComponent } from './reports/myplanet/logs.component';
 
 const routes: Routes = [
   { path: '', component: ManagerDashboardComponent },
