+++ src/app/manager-dashboard/reports/myplanet/logs.component.html
