+++ src/app/manager-dashboard/reports/myplanet/reports.component.html
@@ -107,7 +107,6 @@
     </mat-toolbar-row>
   </ng-template>
 </mat-toolbar>
-
 <div class="space-container">
   <mat-toolbar>
     <mat-toolbar-row class="primary-color font-size-1">
