+++ src/app/manager-dashboard/reports/myplanet/table.component.ts
@@ -7,21 +7,18 @@ import { DialogsViewComponent } from '../../../shared/dialogs/dialogs-view.compo
 
 @Component({
   selector: 'planet-myplanet-table',
-  templateUrl: './myplanet-table.component.html'
+  templateUrl: './table.component.html'
 })
 export class MyPlanetTableComponent implements OnInit, OnChanges, AfterViewInit {
 
   @Input() data = [];
   @Input() dataType: 'logs' | 'report' = 'report';
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatSort) sort: MatSort;
   myPlanets = new MatTableDataSource();
   displayedColumns = [ 'id', 'name' ];
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
-
-  constructor(
-    private dialog: MatDialog,
-    ) {}
+  constructor(private dialog: MatDialog) {}
 
   ngOnInit() {
     this.myPlanets.sortingDataAccessor = (item: any, property) => {
